requirements.txt
pdfplumber
pdfminer.six
